+++ common/common.go
@@ -25,7 +25,8 @@ const (
 	// DefaultDexServerAddr is the HTTP address of the Dex OIDC server, which we run a reverse proxy against
 	DefaultDexServerAddr = "argocd-dex-server:5556"
 	// DefaultRedisAddr is the default redis address
-	DefaultRedisAddr = "argocd-redis:6379"
+	DefaultRedisAddr         = "argocd-redis:6379"
+	DefaultSourcesServerAddr = "sources-server:8090"
 )
 
 // Kubernetes ConfigMap and Secret resource names which hold Argo CD settings