+++ event_reporter/reporter/app_revision_test.go
@@ -56,6 +56,8 @@ func TestGetRevisionsDetails(t *testing.T) {
 			&metrics.MetricsServer{},
 			fakeArgoDb(),
 			"0.0.1",
+			false,
+			nil,
 		}
 
 		result, _ := reporter.getRevisionsDetails(context.Background(), &app, []string{expectedRevision})
@@ -122,6 +124,8 @@ func TestGetRevisionsDetails(t *testing.T) {
 			&metrics.MetricsServer{},
 			fakeArgoDb(),
 			"0.0.1",
+			false,
+			nil,
 		}
 
 		result, _ := reporter.getRevisionsDetails(context.Background(), &app, []string{expectedRevision1, expectedRevision2})
@@ -165,6 +169,8 @@ func TestGetRevisionsDetails(t *testing.T) {
 			&metrics.MetricsServer{},
 			fakeArgoDb(),
 			"0.0.1",
+			false,
+			nil,
 		}
 
 		result, _ := reporter.getRevisionsDetails(context.Background(), &app, []string{expectedRevision})